+++ acme/Types.py
@@ -308,6 +308,7 @@ class ResponseCode(IntEnum):
 	securityAssociationRequired					= 4107
 	invalidChildResourceType					= 4108
 	groupMemberTypeInconsistent					= 4110
+	originatorHasAlreadyRegistered				= 4117
 	appRuleValidationFailed						= 4126
 	internalServerError							= 5000
 	notImplemented								= 5001
@@ -356,6 +357,7 @@ ResponseCode._httpStatusCodes = {											# type: ignore
 		ResponseCode.securityAssociationRequired				: 403,		# SECURITY ASSOCIATION REQUIRED
 		ResponseCode.subscriptionCreatorHasNoPrivilege			: 403,		# SUBSCRIPTION CREATOR HAS NO PRIVILEGE
 		ResponseCode.subscriptionHostHasNoPrivilege				: 403,		# SUBSCRIPTION HOST HAS NO PRIVILEGE
+		ResponseCode.originatorHasAlreadyRegistered				: 403,		# ORIGINATOR HAS ALREADY REGISTERED
 		ResponseCode.appRuleValidationFailed					: 403,		# APP RULE VALIDATION FAILED
 		ResponseCode.notFound									: 404,		# NOT FOUND
 		ResponseCode.operationNotAllowed						: 405,		# OPERATION NOT ALLOWED