+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- [CSE] Corrected response status codes for AE registration errors.
 
 
 ## [0.7.3] - 2021-03-26